requirements.txt
datasets==2.20.0
python-dotenv==1.0.1
psycopg[binary]>=3.1.0
pandas